maintenance: extract client component into /modules folder
